Fixed some annoying typos.
Hook_user contained a small error on #categories: return value should be an array.
Hook_user contained obsolete code that came after a return. removed
Hook_user contained a small error on #categories: the name is an ID and should not be t-ed.

FAPI checkboxes was written as #checkboxs (lost an 'e').

Signed-off-by: Marco Vito Moscaritolo <mavimo@gmail.com>